# Conflicts:
#	interfaces/src/main/java/ar/edu/itba/paw/persistence/BookingDao.java
#	models/src/main/java/ar/edu/itba/paw/models/Booking.java
#	persistence/src/main/resources/alter_prod.sql
#	services/src/main/java/ar/edu/itba/paw/services/ClientServiceImpl.java

+++ interfaces/src/main/java/ar/edu/itba/paw/persistence/ImageDao.java
@@ -3,19 +3,19 @@ package ar.edu.itba.paw.persistence;
 import ar.edu.itba.paw.models.Image;
 
 public interface ImageDao {
-    Image getImage(int id);
+    Image getImage(long id);
 
     Integer uploadImage(String fileName, byte[] imgData);
 
-    Image getpfp(int userId);
+    Image getpfp(long userId);
 
-    Image getVehicleImage(int vehicleId);
+    Image getVehicleImage(long vehicleId);
 
-    Image getPop(int bookingId);
+    Image getPop(long bookingId);
 
-    int uploadPop(byte[] bin, String fileName, int bookingId);
+    long uploadPop(byte[] bin, String fileName, long bookingId);
 
-    int uploadVehicleImage(byte[] bin, String fileName, int vehicleId);
+    long uploadVehicleImage(byte[] bin, String fileName, long vehicleId);
 
-    int uploadPfp(byte[] bin, String fileName, int userId);
+    long uploadPfp(byte[] bin, String fileName, long userId);
 }
