+++ models/src/main/java/ar/edu/itba/paw/models/Image.java
@@ -1,17 +1,17 @@
 package ar.edu.itba.paw.models;
 
 public class Image {
-    private final int id;
+    private final long id;
     private final byte[] data;
     private final String fileName;
 
-    public Image(int id, byte[] data, String fileName) {
+    public Image(long id, byte[] data, String fileName) {
         this.id = id;
         this.data = data;
         this.fileName = fileName;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
